fix: adding a user comment field to the model
