+++ linear_api/domain/user_models.py
@@ -178,3 +178,22 @@ class Reaction(LinearModel):
     emoji: str
     user: Optional[LinearUserReference] = None
     createdAt: datetime
+
+
+class Comment(LinearModel):
+    """Represents a comment in Linear"""
+    linear_class_name: ClassVar[str] = "Comment"
+
+    id: str
+    body: str
+    user: LinearUserReference
+    createdAt: datetime
+    updatedAt: datetime
+
+
+class CommentConnection(LinearModel):
+    """Connection model for comments"""
+    linear_class_name: ClassVar[str] = "CommentConnection"
+
+    nodes: List[Comment] = Field(default_factory=list)
+    pageInfo: Optional[Dict[str, Any]] = None
